feat: LSP signature help for assert and assert_eq (#5862)
# Description

## Problem

This is something minor, but it's good for completeness and also for
users to easily learn that there's an optional failure message you can
use in these built-in functions.

## Summary


![lsp-signature-help-assert](https://github.com/user-attachments/assets/77d621fb-360f-47df-92da-f51af00298db)


![lsp-signature-help-assert_eq](https://github.com/user-attachments/assets/6d23a051-9440-4a86-b970-e060b60ab5a1)

## Additional Context

I don't know if the syntax is clear for showing that the last argument
is optional, so if you have another syntax in mind let me know!

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.