fix: make simulations validate resulting tx by default (#8157)
Closes #7956.

This PR simply flips a switch to make all simulations fail by default if
the transaction is an invalid one. Previously it did not throw due to
wanting to minimize disruptions in the PR, but it seems like there were
not that many effects. The only exception is the e2e prover. These are
failing due to a prover verification, and I have made an issue for this
to be looked into.