feat(perf): Simplify poseidon2 cache zero-pad (#5869)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Works towards general Brillig bytecode size reductions

## Summary\*

When actually setting up the state for a poseidon2 permutation, we have
a loop to zero-pad the cache and then another loop to add the cache to
the state. If we just have a single loop that only adds to the state
when the rate index is below the cache size we achieve the same goal
with less code.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.