feat: add `FunctionDef::has_named_attribute` (#5870)
# Description

## Problem

Part of #5668

## Summary

Adds `FunctionDef::has_named_attributes` and also changes
`Expr::resolve` to take an optional `FunctionDefinition` so you can
resolve it relative to another function.

## Additional Context

Now the `inject_context` macro works like it needs to for
Aztec-Packages!... with the caveat that not all expressions are
currently handled in `Expr::mutate` yet.

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>