feat: LSP code actions to import or qualify unresolved paths (#5876)
# Description

## Problem

Part of #1579

## Summary

There are many code actions we could support, but the most useful ones
are importing unknown identifiers or fully-qualifying them. This is what
this PR is about.


![lsp-code-actions](https://github.com/user-attachments/assets/6107bb2b-ed2d-4430-a213-5a3d4a6e1388)

Also thanks to VS Code that it notices that if you request code action
on an error, it actually requests code actions for that error span... so
it seems like the two are connected, but they are not! And it works just
fine :-)

Also fixes a bug I found in "go to definition" when there was no
definition to resolve to (precisely the case this code action works on).

## Additional Context

Note that in Rust Analyzer it shows as "Qualify [name]" or "Import
[name]" and if there are multiple possibilities, a popup shows up
letting you choose which one. I wanted to do the same, but looking at
the server trace for Rust Analyzer I see each code action has a "group"
property, and that's how they are grouped, but... I couldn't find that
property in the lsp-types crate, or even in the LSP specification 😮 . So
for now each possibility will show up (no intermediate popup). But I
think that's better than not having this functionality at all, and we
could always improve it later if we find out how.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>