feat: LSP diagnostics for all package files (#5895)
# Description

## Problem

I noticed that currently LSP will only produce diagnostics for the file
you save. That's not very convenient, as if for example you make a
change to a function signature that breaks someone using it, you won't
find out unless you go to that other file and save it (but how do you
know which file is it?) My guess is that you'd use `nargo` on the
command line to find out.

## Summary

Now when a package is type-checked, all errors are reported, regardless
of whether they match the file that was saved. I believe this will speed
up developing with Nargo! (it was also slowing me down when trying out
new warnings or errors)


![lsp-error-on-all-files](https://github.com/user-attachments/assets/2441e1dd-15fc-4f23-826c-7ba833012c0e)

## Additional Context

I thought implementing this was going to be straight-forward (just start
reporting errors in all files) but there's a catch: if a file had errors
but doesn't have errors anymore, we still have to notify that the file
now **doesn't** have errors. This requires us tracking which files
currently have errors and making a diff with the new errors.

That said: maybe there was a reason errors were only reported in the
file that was saved?

Final comment: ideally when you save a file in a package, all packages
that depend on this package would also get type-checked... but that's
currently slow (or noticeably slow), but I hope we could eventually make
it work (like in Rust Analyzer).

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.