feat(perf): Remove last store in return block if last load is before that store (#5910)
# Description

## Problem\*

Working towards #4535 

## Summary\*

Small optimization found while working in mem2reg. Now that we have a
small per function state we can see which stores/loads that are leftover
can be removed.

A place we are potentially missing stores that can be removed is in
return blocks. If we have a store inside of a return block that is not
loaded from inside that block, we can safely remove that store (given it
is not a reference param).

## Additional Context


## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: TomAFrench <tom@tomfren.ch>