feat: unquote some value as tokens, not as unquote markers (#5924)
# Description

## Problem

Resolves #5916

## Summary

## Additional Context

I think there might be other `Value`s that we could convert into tokens,
like `String`, but I didn't want to handle those in this PR (should it
be `Str` or `RawStr`? etc.). And it's possible that we only need this
special logic for integers anyway.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.