chore: error on false constraint (#5890)
# Description

## Problem\*

Report a bug during compilation when a constraint is false

## Summary\*

It is done in acir-gen which ensures that all side effects are resolved.


## Additional Context

The reported bug does not fail the compilation, because it will mess
around compile/run-time failures and also for the Noir tests. We want a
consistent handling of assert, whether it fails at compile-time,
run-time, or test-time.

## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>