feat: Add `StructDefinition::set_fields` (#5931)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5911

## Summary\*

Adds a function to set the fields on a struct type

## Additional Context

Hyper-specific error messages can be fun. Here's the error issued when
one of the field names isn't a valid identifier:

```
error: Quoted value in index 1 of this slice is not a valid field name
   ┌─ src/main.nr:25:18
   │
25 │     s.set_fields(fields);
   │                  ------ `quote { foo bar }` is not a valid field name for `set_fields`
   │
```

## Documentation\*

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Ary Borenszweig <asterite@gmail.com>