feat: add `StructDefinition::add_attribute` and `has_named_attribute` (#5945)
# Description

## Problem

Resolves #5874

## Summary

## Additional Context

I originally wanted to test this by adding `abi(...)` to a struct and
seeing that it errors if that struct appears in a signature but is
outside of a contract, but... function parameters are type-checked
before that (I think). So another way to test this is to also have
`has_named_attribute`, so I added that too.

Side note: maybe adding an attribute to a function/struct should
eventually invoke the associated function, if there's any, but we can
probably try to do that in a separate PR, and only if really needed.

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.