feat: LSP now suggests self fields and methods (#5955)
# Description

## Problem

While working with Rust I noticed that if you type a field name or a
method name and there's a `self` variable, those fields and methods are
suggested without you having to explicitly write `self`. I think that's
useful as you sometimes forget about the self prefix, or just because
you end up typing less.

## Summary

This PR implements that.


![lsp-self](https://github.com/user-attachments/assets/976151ab-cdb2-4690-8b62-21514314229a)

## Additional Context



## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.