fix(mem2reg): Handle aliases better when setting a known value for a load (#5959)
# Description

## Problem\*

Resolves #5771 

This was a bug found on
https://github.com/AztecProtocol/aztec-packages/pull/8378 from
https://github.com/noir-lang/noir/pull/5935. However it looks to
inadvertently fix the linked issue as well.

## Summary\*

We were just directly inserting a new expression and alias for a load
result. This was overriding whatever expression of AliasSet may have
been there before. This PR switches to checking whether the result
already has an expression, which if it does to use that. It then checks
whether the result already has an alias set, if it does we add to the
alias set rather than overriding it.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.