feat: add a `panic` method to the stdlib (#5966)
# Description

## Problem

Resolves #5849

## Summary

Adds a `panic` method exactly as proposed in #5849 , and uses it in a
couple of places where `panic` would have been useful to have.

## Additional Context

Is it okay in `std::panic::panic`? Where should we document this? (maybe
in a new `panic.md` file, or somewhere else?)

## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>