feat: let `nargo` and LSP work well in the stdlib (#5969)
# Description

## Problem

`nargo test` doesn't work in the standard library, and LSP gives lots of
(incorrect) errors when opening the standard library.

## Summary

The main issue is that when you run `nargo` against the standard
library, a copy of it is also loaded as the standard library. Then we
get duplicate definitions, etc.

I tried several things to solve this and one that worked and ended up
making a bit of sense is having a crate be the root and the standard
library at the same time.

Now with this PR we can run `nargo test`, LSP works well, and you can
even run tests directly from the editor, which could speed up the stdlib
development.

## Additional Context

1. 7 warnings still remain in the standard library (all unused
functions) but I'd like to solve that in a separate PR.
2. If this gets merged we could eventually change CI to compile nargo,
then run `nargo test`, instead of running stdlib tests as part of one
crate's tests.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.