chore: remove 3 unused functions warnings in the stdlib (#5973)
# Description

## Problem

Two functions were unused, another one two but it was kind of a test so
I marked it as such.

## Summary


## Additional Context

After this 4 "unused" functions remain, but solving those is a bit more
challenging so I'm leaving that for a separate PR.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.