feat: add `Expr::as_let` (#5964)
# Description

## Problem

Part of #5668

## Summary

Adds `Expr::as_let`, but also introduces a new `ExprValue` for patterns,
their inlining code, etc.

## Additional Context

I wonder if `Pattern` should be an `Expr` or something else, because it
can't be converted to an expression like a statement or like an LValue 🤔

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.