chore: document BoundedVec (#5974)
# Description

## Problem

Part of #5797

## Summary

This PR copies the docs from the `.md` file to the `.nr` file.

## Additional Context

I'll be sending one PR per `.md` file. It looks like they are 36 in
total so it will take some time to port all the docs, but not that much.
And this way reviewing these PRs will be more manageable.

For now examples are copied inline into the `.nr` file. Eventually we
could maybe have `nargo test` run or at least type-check these example

## Documentation

Check one:
- [ ] No documentation needed.
- [x] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.