fix: Error when `quote` is used in runtime code (#5978)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Fixes a few small issues:
- Error when `quote` is used in runtime code
  - Had to update some stdlib code for this
- This could maybe be an issue in the future for something like `impl
Append for Quoted` which relies on this. These impl methods have to be
comptime now which we currently allow even though the parent trait says
nothing about it. At least with
https://github.com/noir-lang/noir/pull/5976 we'd get an error if one of
these accidentally sneaks into runtime code.
- `self.in_comptime_context()` wasn't being tracked accurately in the
case of attributes and globals where the function context would only be
of length 2 due to there being no outer function. I changed the check to
make it more direct instead of relying on indirect changes.
- Comptime globals weren't evaluated in their own function context
before so we waited to resolve traits until the end of the program. This
could lead to trait errors when interpreting them.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [ ] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.