feat: LSP completion function detail (#5993)
# Description

## Problem

Just a tiny thing: when functions are suggested, Rust Analyzer shows the
full function signature above the documentation (this is a completion
item's `detail` property). It's useful because in the list the signature
is usually trimmed.

## Summary

Before:


![image](https://github.com/user-attachments/assets/173acca4-74fe-40ff-9c8c-6552296c05e4)

After:


![image](https://github.com/user-attachments/assets/520fc7df-bf01-4ef2-887a-b4e651f4035c)

## Additional Context

Also includes a bunch of small refactors to make it easier to test that
something is a function completion item regardless of whether it ends up
being a snippet completion or not, whether it triggers a command, etc.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.